backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
pydantic==2.6.1
python-dotenv==1.0.0
